Move files, rename various exports to make a little more sense
