+++ src/lib.rs
@@ -1,0 +1,1 @@
+pub mod st7701s_spi;
